bilinear_interpolation_scaling/requirements.txt
pillow==10.0.0
numpy==1.26.0